fix: validate jwt

+++ scripts/restart_services.bash
@@ -5,12 +5,6 @@ eval $(minikube -p minikube docker-env)
 
 ID=$(docker images --format "{{.ID}} {{.Repository}} {{.Tag}}" | grep "taxi/$svcname latest" | awk '{print $1}')
 
-if [ -z "$ID" ]
-then
-    echo "Image not found"
-    exit 1
-fi
-
 if [ "$svcname" != "usermgmt" ]
 then
     make build-$svcname
@@ -26,5 +20,11 @@ kubectl rollout restart deployment $svcname
 kubectl rollout status deployment $svcname
 
 sleep 60
+
+if [ -z "$ID" ]
+then
+    echo "No image found"
+    exit 0
+fi
 
 docker rmi -f $ID
